Refactor ScreenshotUtils into element / ExtWebDriver methods, rename class to ScreenshotComparator and include only compare methods Conflicts: 	src/main/java/org/finra/jtaf/ewd/utils/ScreenshotUtils.java
